requirements.txt
Pillow==8.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
